requirements.txt
colorama==0.4.6
mysql-connector-python==8.4.0
pillow==10.3.0
pypng==0.20220715.0
python-barcode==0.15.1
qrcode==7.4.2
tk==0.1.0
typing_extensions==4.12.0
